+++ rust/blockstore/src/arrow/block/delta/mod.rs
@@ -1,5 +1,7 @@
+pub(super) mod data_record;
 mod delta;
 pub(super) mod int32;
+pub(super) mod roaring_bitmap;
 mod storage;
 pub(super) mod string;
 pub(super) mod uint32;
